requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-dotenv
colorama
pytest
pytest-mock